requirements.txt
asgiref==3.5.0
autopep8==1.6.0
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
